CarbonTakeOrderedAndProjectExec changes
